Add Middleware

+++ app/Http/Controllers/api/TaskController.php
@@ -2,12 +2,15 @@
 
 namespace App\Http\Controllers\api;
 use App\Http\Controllers\Controller;
-
+use App\Models\Task;
+use App\Policies\TaskPolicy;
 use App\Services\TaskService;
 use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class TaskController extends Controller
 {
+    use AuthorizesRequests;
     protected $taskService;
 
     public function __construct(TaskService $taskService)
@@ -20,6 +23,7 @@ class TaskController extends Controller
      */
     public function index()
     {
+      
         return response()->json($this->taskService->getAllTasks());
     }
 
@@ -28,6 +32,7 @@ class TaskController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Task::class);
         $task = $this->taskService->createTask($request->all());
         return response()->json($task, 201);
     }
@@ -37,6 +42,7 @@ class TaskController extends Controller
      */
     public function show(string $id)
     {
+        
         return response()->json($this->taskService->getTaskById($id));
     }
 
@@ -45,6 +51,7 @@ class TaskController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('update', Task::class);
         $this->taskService->updateTask($id, $request->all());
         return response()->json(['message' => 'Task updated successfully']);
     }
@@ -54,6 +61,7 @@ class TaskController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete', Task::class);
         $this->taskService->deleteTask($id);
         return response()->json(['message' => 'Task deleted successfully']);
     }
